pyproject.toml
[tool.yapf]
based_on_style = "pep8"
column_limit = 88
indent_width = 4
use_tabs = false
blank_line_before_nested_class_or_def = true
split_before_logical_operator = true
split_before_arithmetic_operator = true

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "meetingez"
version = "0.1.0"
description = "智能会议助手 - 基于 OpenAI Realtime API 的实时语音识别和翻译工具"
authors = [{name = "MeetingEZ Team"}]
license = {text = "MIT"}
readme = "docs/README.md"
requires-python = ">=3.8"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Communications :: Conferencing",
    "Topic :: Multimedia :: Sound/Audio :: Speech",
]
dependencies = [
    "Flask>=2.3.0",
    "Werkzeug>=2.3.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]
